converted Header and Footer to ViewComponents
