+++ src/Submissions/Api/Views/Shared/Components/Footer/Default.cshtml
@@ -1,4 +1,4 @@
-﻿@using System.Configuration
+@using System.Configuration
 @using Biobanks.Submissions.Api.Models.Footer
 @model FooterModel
